requirements.txt
Flask  
gunicorn
tensorflow
opencv-python
numpy
Pillow